requirements-minimal.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-dotenv>=0.19.0
python-multipart>=0.0.5
email-validator>=1.1.3
aiosmtplib>=1.1.6
